+++ BackEnd/src/main/java/com/portfolio/BackEnd/Repository/IRedesSocialesRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IRedesSocialesRepository extends JpaRepository<RedesSociales, Long> {
+public interface IRedesSocialesRepository extends JpaRepository<RedesSociales, Integer> {
     
 }
